LocalOCR/requirements.txt
rumps>=0.4.0
flask>=2.3.0
rapidocr-onnxruntime>=1.3.0
pillow>=9.0.0
werkzeug>=2.3.0
pywebview>=4.0.0
pyobjc-framework-ApplicationServices
